+++ src/main/java/com/hospital/management/entities/State.java
@@ -14,7 +14,7 @@ public class State {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name= "STATE_ID")
-    private  Integer stateId;
+    private  Long stateId;
 
     @Column(name="STATE_CODE")
     private  String stateCode;
